pyproject.toml
[tool.poetry]
name = "team-cac-discord-bot"
version = "0.1.0"
description = "discord bot for the team cac discord server"
authors = ["peter <peterbuoy@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
"discord.py" = "^1.7.3"
python-dotenv = "^0.20.0"
pytz = "^2022.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
